requirements.txt
docopt==0.6.2
PyYAML==3.12
RPi.GPIO==0.6.3
